@@ -6,20 +6,15 @@
     "build": "vite build"
   },
   "dependencies": {
-    "@ant-design/colors": "^4.0.5",
-    "ant-design-vue": "^2.0.0-beta.10",
-    "vue": "^3.0.0-rc.1",
-    "vue-router": "^4.0.0-beta.13",
-    "vuex": "^4.0.0-beta.4"
+    "vue": "^3.0.5",
+    "vue-router": "^4.0.3"
   },
   "devDependencies": {
-    "@vue/compiler-sfc": "^3.0.0-rc.1",
-    "babel-plugin-import": "^1.13.1",
-    "eslint": "^7.11.0",
-    "eslint-plugin-vue": "^7.1.0",
-    "prettier": "^2.1.2",
-    "sass": "^1.27.0",
-    "typescript": "^4.0.3",
-    "vite": "^1.0.0-rc.1"
+    "@types/node": "^14.14.22",
+    "@vitejs/plugin-vue": "^1.0.4",
+    "@vue/compiler-sfc": "^3.0.5",
+    "sass": "^1.32.5",
+    "typescript": "^4.1.3",
+    "vite": "^2.0.0-beta.12"
   }
 }
