@@ -3,22 +3,17 @@ import { createRouter, createWebHistory } from 'vue-router'
 const routes = [
   {
     path: '/',
-    name: 'Home',
-    component: () => import('/@/views/Home.vue')
+    name: 'home',
+    component: () => import('../views/Home.vue')
   },
   {
-    path: '/setting',
-    name: 'Setting',
-    component: () => import('/@/views/Setting.vue')
-  },
-  {
-    path: '/lifeCycle',
-    name: 'lifeCycle',
-    component: () => import('/@/views/LifeCycle.vue')
+    path: '/base',
+    name: 'base',
+    component: () => import('../views/Base.vue')
   }
 ]
 
 export default createRouter({
-  history: createWebHistory('/krry/'),
+  history: createWebHistory('/'),
   routes
 })
