@@ -1,17 +1,11 @@
 <template>
-  <h1>This is a home page</h1>
-  <HelloWorld msg="Hello Vue 3.0 + Vite" />
+  <hello-world></hello-world>
 </template>
 
-<script lang="ts">
-import { defineComponent } from 'vue'
-
+<script setup>
 import HelloWorld from '../components/HelloWorld.vue'
-
-export default defineComponent({
-  name: 'Home',
-  components: {
-    HelloWorld
-  }
-})
 </script>
+
+<style>
+
+</style>
